requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
sentence-transformers
chromadb
langchain
transformers
torch
gradio
streamlit
jupyter

langchain-community
langchain-chroma
faiss-cpu
huggingface-hub
accelerate
protobuf
tiktoken
python-dotenv

pytest
pytest-cov
pytest-mock
coverage
psutil